feat: make max nb matches per string a u32
This allows to generate a Value during evaluation safely, and avoid
some unreachable poison case.

+++ boreal/src/evaluator/ac_scan.rs
@@ -192,7 +192,7 @@ impl AcScan {
         };
 
         if let AcResult::Matches(matches) = &mut matches[variable_index] {
-            matches.truncate(params.string_max_nb_matches);
+            matches.truncate(params.string_max_nb_matches as usize);
         }
     }
 }
